feat: 도서 등록 기능 테스트 작성

+++ src/main/java/shop/nuribooks/books/entity/Reviews.java
@@ -36,7 +36,7 @@ public class Reviews {
 	@ColumnDefault("0")
 	@Column(nullable = false)
 	//@Column(nullable = false, columnDefinition = "tinyint(1)")
-	private byte score; //평점 1-5점
+	private int score; //평점 1-5점
 
 	@NotNull
 	private LocalDateTime createdAt;
